+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - WIP
+
+- GSQ CountTrue: meta query that checks at least X GSQs are true.
+- Expanded StoredQuery with some machine fields, can't do EMC type things though.
+
 ## [1.0.0] - 2025-01-25
 
 ### Added